Master Projects/IoT - Load Shedding System/Implementation/requirements.txt
apturl==0.5.2
asttokens==2.4.1
attrs==23.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
CoAPthon3==1.0.2
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
crcmod==1.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
ecdsa==0.18.0
emlearn==0.20.4
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
fonttools==4.53.0
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
intelhex==2.3.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libusb1==1.9.3
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
netifaces==0.11.0
nrfutil==6.1.7
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pc-ble-driver-py==0.17.0
pexpect==4.8.0
piccata==2.0.3
Pillow==9.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.12.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
pyspinel==1.0.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
referencing==0.35.1
reportlab==3.6.8
requests==2.25.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
SecretStorage==3.3.1
six==1.16.0
stack-data==0.6.3
systemd-python==234
tabulate==0.9.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
wrapt==1.16.0
xdg==5
xgboost==2.0.3
xkit==0.0.0
zipp==1.0.0